Add eval reorg + tests

+++ src/assistant/control/AssistantCtrlPlane.ts
@@ -2,6 +2,7 @@ import {
   CreateAssistantRequestRegionEnum,
   ManageAssistantsApi as ManageAssistantsApiCtrl,
 } from '../../pinecone-generated-ts-fetch/assistant_control';
+import { MetricsApi } from '../../pinecone-generated-ts-fetch/assistant_evaluation';
 import { AssistantHostSingleton } from '../assistantHostSingleton';
 import { PineconeConfiguration } from '../../data';
 import {
@@ -12,25 +13,41 @@ import { deleteAssistantClosed } from './deleteAssistant';
 import { getAssistantClosed } from './getAssistant';
 import { listAssistantsClosed } from './listAssistants';
 import { updateAssistant, updateAssistantClosed } from './updateAssistant';
+import { Eval, evaluateClosed } from './evaluate';
 
 export class AssistantCtrlPlane {
-  _createAssistant: ReturnType<typeof createAssistantClosed>;
-  _deleteAssistant: ReturnType<typeof deleteAssistantClosed>;
-  _getAssistant: ReturnType<typeof getAssistantClosed>;
-  _listAssistants: ReturnType<typeof listAssistantsClosed>;
-  _updateAssistant: ReturnType<typeof updateAssistantClosed>;
+  readonly _createAssistant?: ReturnType<typeof createAssistantClosed>;
+  readonly _deleteAssistant?: ReturnType<typeof deleteAssistantClosed>;
+  readonly _getAssistant?: ReturnType<typeof getAssistantClosed>;
+  readonly _listAssistants?: ReturnType<typeof listAssistantsClosed>;
+  readonly _updateAssistant?: ReturnType<typeof updateAssistantClosed>;
+  readonly evalApi?: MetricsApi;
+  readonly _evaluate?: ReturnType<typeof evaluateClosed>;
+
   private config: PineconeConfiguration;
 
-  constructor(api: ManageAssistantsApiCtrl, config: PineconeConfiguration) {
+  constructor(
+    config: PineconeConfiguration,
+    params: { assistantApi?: ManageAssistantsApiCtrl; evalApi?: MetricsApi }
+  ) {
     this.config = config;
-    this._createAssistant = createAssistantClosed(api);
-    this._deleteAssistant = deleteAssistantClosed(api);
-    this._getAssistant = getAssistantClosed(api);
-    this._listAssistants = listAssistantsClosed(api);
-    this._updateAssistant = updateAssistantClosed(api);
+    if (params.evalApi) {
+      this.evalApi = params.evalApi;
+      this._evaluate = evaluateClosed(this.evalApi);
+    }
+    if (params.assistantApi) {
+      this._createAssistant = createAssistantClosed(params.assistantApi);
+      this._deleteAssistant = deleteAssistantClosed(params.assistantApi);
+      this._getAssistant = getAssistantClosed(params.assistantApi);
+      this._listAssistants = listAssistantsClosed(params.assistantApi);
+      this._updateAssistant = updateAssistantClosed(params.assistantApi);
+    }
   }
 
   async createAssistant(options: createAssistantRequest) {
+    if (!this._createAssistant) {
+      throw new Error('Assistant API is not initialized');
+    }
     if (options.region) {
       let region: CreateAssistantRequestRegionEnum =
         CreateAssistantRequestRegionEnum.Us;
@@ -46,28 +63,46 @@ export class AssistantCtrlPlane {
         region =
           options.region.toLowerCase() as CreateAssistantRequestRegionEnum;
       }
-
       AssistantHostSingleton._set(this.config, options.name, region);
-
       options.region = region;
     }
-
     return await this._createAssistant(options);
   }
 
   async deleteAssistant(assistantName: string) {
+    if (!this._deleteAssistant) {
+      throw new Error('Assistant API is not initialized');
+    }
     return await this._deleteAssistant(assistantName);
   }
 
   async getAssistant(assistantName: string) {
+    if (!this._getAssistant) {
+      throw new Error('Assistant API is not initialized');
+    }
     return await this._getAssistant(assistantName);
   }
 
   async listAssistants() {
+    if (!this._listAssistants) {
+      throw new Error('Assistant API is not initialized');
+    }
     return await this._listAssistants();
   }
 
   async updateAssistant(options: updateAssistant) {
+    if (!this._updateAssistant) {
+      throw new Error('Assistant API is not initialized');
+    }
     return await this._updateAssistant(options);
+  }
+
+  // Eval is currently spec-d to be on the data plane, but it lives here since you don't need an assistantName to
+  // access it, as you do for the other data plane operations
+  evaluate(options: Eval) {
+    if (!this._evaluate) {
+      throw new Error('Evaluation API is not initialized');
+    }
+    return this._evaluate(options);
   }
 }
